Add aziot-keyd and aziot-certd OpenAPI specs.
Other changes:

- Fix naming of `preferredAlgorithms` field of keyd's
  create-key-pair-if-not-exists API to use camel case instead of snake case.

- Merge certd's import-cert API into the get-or-delete-cert API so that
  the import-cert API can actually be invoked. Previously it would never get
  invoked because the get-or-delete-cert route would match the same URLs
  and end up returning HTTP 405.
